main - minimal example
