+++ Cargo.toml
@@ -1,15 +1,16 @@
 [package]
 name = "rtfm-app"
-version = "0.1.2"
+version = "0.2.0"
+description = "RFTM - v0.4beta"
 authors = ["Cecília Carneiro <cecilia.carneiroesilva@gmail.com>"]
 
 
 [dependencies]
-cortex-m = { git = "https://github.com/ceciliacsilva/cortex-m", version = "0.5.7" }
-cortex-m-rt = {version = "0.6.4", features = ["device"]} #feature device
-cortex-m-rtfm = { git = "https://github.com/ceciliacsilva/cortex-m-rtfm" }
+cortex-m = "0.5.8"
+cortex-m-rt = { version = "0.6.5", features = ["device"] }
+cortex-m-rtfm = "0.4.0-beta.2"
 cortex-m-semihosting = "0.3.1"
-hal = { path = "../quickstart/narc_rust/hal", version = "0.1.5" }
+narc-hal = { path = "../quickstart/narc_rust/hal", version = "0.1.5" }
 heapless = { version = "0.4.0", features = ["const-fn"] }
 nb = "0.1.1"
